+++ src/gemseo_benchmark/problems/problems_group.py
@@ -113,6 +113,7 @@ class ProblemsGroup:
         max_eval_number: int = 0,
         plot_kwargs: Mapping[str, ConfigurationPlotOptions] = READ_ONLY_EMPTY_DICT,
         grid_kwargs: Mapping[str, str] = READ_ONLY_EMPTY_DICT,
+        use_evaluation_log_scale: bool = False,
     ) -> None:
         """Generate the data profiles of given algorithms relative to the problems.
 
@@ -128,6 +129,8 @@ class ProblemsGroup:
             plot_kwargs: The keyword arguments of `matplotlib.axes.Axes.plot`
                 for each algorithm configuration.
             grid_kwargs: The keyword arguments of `matplotlib.pyplot.grid`.
+            use_evaluation_log_scale: Whether to use a logarithmic scale
+                for the number of function evaluations axis.
         """
         # Initialize the data profile
         target_values = {
@@ -158,6 +161,7 @@ class ProblemsGroup:
             file_path=plot_path,
             plot_kwargs=plot_kwargs,
             grid_kwargs=grid_kwargs,
+            use_evaluation_log_scale=use_evaluation_log_scale,
         )
 
     def __len__(self) -> int: