+++ src/gemseo_benchmark/report/report.py
@@ -153,6 +153,7 @@ class Report:
         use_log_scale: bool = False,
         plot_only_median: bool = False,
         use_time_log_scale: bool = False,
+        use_evaluation_log_scale: bool = False,
     ) -> None:
         """Generate the benchmarking report.
 
@@ -165,6 +166,8 @@ class Report:
             plot_only_median: Whether to plot only the median and no other centile.
             use_time_log_scale: Whether to use a logarithmic scale
                 for the time axis.
+            use_evaluation_log_scale: Whether to use a logarithmic scale
+                for the number of function evaluations axis.
         """
         self.__create_root_directory()
         self.__create_algos_file()
@@ -175,6 +178,7 @@ class Report:
             use_log_scale,
             plot_only_median,
             use_time_log_scale,
+            use_evaluation_log_scale,
         )
         self.__create_index()
         self.__build_report(to_html, to_pdf)
@@ -268,6 +272,7 @@ class Report:
         use_log_scale: bool = False,
         plot_only_median: bool = False,
         use_time_log_scale: bool = False,
+        use_evaluation_log_scale: bool = False,
     ) -> None:
         """Create the files corresponding to the benchmarking results.
 
@@ -278,6 +283,8 @@ class Report:
             plot_only_median: Whether to plot only the median and no other centile.
             use_time_log_scale: Whether to use a logarithmic scale
                 for the time axis.
+            use_evaluation_log_scale: Whether to use a logarithmic scale
+                for the number of function evaluations axis.
         """
         self.__fill_template(
             self.__root_directory / FileName.RESULTS.value,
@@ -290,6 +297,7 @@ class Report:
                     use_log_scale,
                     plot_only_median,
                     use_time_log_scale,
+                    use_evaluation_log_scale,
                 )
                 for group in self.__algorithms_configurations_groups
             ],
@@ -303,6 +311,7 @@ class Report:
         use_log_scale: bool,
         plot_only_median: bool,
         use_time_log_scale: bool,
+        use_evaluation_log_scale: bool,
     ) -> str:
         """Create the results files of a group of algorithm configurations.
 
@@ -314,6 +323,8 @@ class Report:
             plot_only_median: Whether to plot only the median and no other centile.
             use_time_log_scale: Whether to use a logarithmic scale
                 for the time axis.
+            use_evaluation_log_scale: Whether to use a logarithmic scale
+                for the number of function evaluations axis.
 
         Returns:
             The path to the main file.
@@ -353,6 +364,7 @@ class Report:
                     use_log_scale,
                     plot_only_median,
                     use_time_log_scale,
+                    use_evaluation_log_scale,
                 )
                 .relative_to(results_root)
                 .as_posix()
@@ -379,6 +391,7 @@ class Report:
         use_log_scale: bool,
         plot_only_median: bool,
         use_time_log_scale: bool,
+        use_evaluation_log_scale: bool,
     ) -> Path:
         """Create the results file of a group of algorithm configurations.
 
@@ -393,6 +406,8 @@ class Report:
             plot_only_median: Whether to plot only the median and no other centile.
             use_time_log_scale: Whether to use a logarithmic scale
                 for the time axis.
+            use_evaluation_log_scale: Whether to use a logarithmic scale
+                for the number of function evaluations axis.
 
         Returns:
             The path to the main file.
@@ -409,7 +424,11 @@ class Report:
             plot_kwargs=self.__plot_kwargs,
         )
         figures = plotter.plot(
-            plot_all_histories, use_log_scale, plot_only_median, use_time_log_scale
+            plot_all_histories,
+            use_log_scale,
+            plot_only_median,
+            use_time_log_scale,
+            use_evaluation_log_scale,
         )
 
         # Create the file dedicated to the group of problems
@@ -425,7 +444,9 @@ class Report:
             ],
             problems_group_name=problems.name,
             problems_group_description=problems.description,
-            data_profile=self.__get_relative_path(plotter.plot_data_profiles()),
+            data_profile=self.__get_relative_path(
+                plotter.plot_data_profiles(use_evaluation_log_scale)
+            ),
             problems_names=[problem.name for problem in problems],
             group_problems_paths=[
                 self.__create_problem_results_files(