+++ tests/data_profiles/test_data_profile.py
@@ -68,16 +68,26 @@ def test_compute_data_profiles():
     assert profiles["algo"] == [0.0, 0.0, 0.5, 0.5, 0.5, 1.0]
 
 
-@image_comparison(
-    baseline_images=["data_profile"], remove_text=True, extensions=["png"]
+@pytest.mark.parametrize(
+    ("baseline_images", "use_evaluation_log_scale"),
+    [
+        (
+            [f"data_profile[use_evaluation_log_scale={use_evaluation_log_scale}]"],
+            use_evaluation_log_scale,
+        )
+        for use_evaluation_log_scale in [False, True]
+    ],
 )
-def test_plot_data_profiles():
+@image_comparison(None, ["png"])
+def test_plot_data_profiles(baseline_images, use_evaluation_log_scale):
     """Check the data profiles figure."""
     data_profile = DataProfile({"problem": TargetValues([1.0, 0.0])})
     data_profile.add_history("problem", "algo", [2.0, 1.5, 1.0, 0.5, 0.1, 0.0])
     data_profiles = data_profile.compute_data_profiles("algo")
     pyplot.close("all")
-    data_profile._plot_data_profiles(data_profiles)
+    data_profile._plot_data_profiles(
+        data_profiles, use_evaluation_log_scale=use_evaluation_log_scale
+    )
 
 
 @pytest.mark.parametrize("converter", [lambda _: _, str])